+++ ProefProeveVeluwe/Assets/Scripts/Runtime/Animal Interaction/Feeding.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3817c81ae7c67974bb269ec69d964748
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
