+++ ProefProeveVeluwe/ProjectSettings/EditorBuildSettings.asset
@@ -9,6 +9,6 @@ EditorBuildSettings:
     path: Assets/Scenes/Test/GPSDebugTest.unity
     guid: 99c9720ab356a0642a771bea13969a05
   - enabled: 1
-    path: Assets/Scenes/Minigames/PettingMinigame.unity
-    guid: 74a0770435262e1449e00eefcef2af95
+    path: Assets/Scenes/Minigames/FeedAnimal.unity
+    guid: 5bcc8a835dcbc3242810819e5e354257
   m_configObjects: {}
