+++ ProefProeveVeluwe/ProjectSettings/TagManager.asset
@@ -5,6 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - Pettable
+  - Food
   layers:
   - Default
   - TransparentFX
